Restructure file-adapters package
