+++ .changeset/f56dc666/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/file-adapters", "type": "patch" }], "dependents": [] }
