+++ .changeset/f56dc666/changes.md
@@ -1,0 +1,1 @@
+- Restructure internal code
