+++ packages/file-adapters/lib/cloudinary.js
